# backend/app/services/migration_advisor/migration_advisor/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Cloud Migration Advisor API
  description: |
    Comprehensive API for the Cloud Migration Advisor platform that guides organizations through
    their cloud migration journey from initial assessment through provider selection, migration
    execution, and post-migration resource organization.
    
    ## Features
    - Migration project management and assessment
    - Workload and requirements analysis
    - ML-powered cloud provider recommendations
    - Migration planning and execution tracking
    - Automated resource discovery and organization
    - Multi-dimensional resource management
    - Post-migration FinOps integration
    
    ## Authentication
    All endpoints require authentication using Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
  version: 1.0.0
  contact:
    name: Cloud Migration Advisor Support
    email: support@cloudmigration.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.cloudmigration.example.com/v1
    description: Production server
  - url: https://staging-api.cloudmigration.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
...
    MigrationPlan:
      type: object
      properties:
        plan_id:
          type: string
          description: Unique plan identifier
        target_provider:
          type: string
          description: Target cloud provider
        total_duration_days:
          type: integer
          description: Total migration duration in days
        estimated_cost:
          type: number
          format: float
          description: Estimated migration cost
        risk_level:
          type: string
          enum: [low, medium, high, critical]
          description: Migration risk level
        phases:
          type: array
          items:
            $ref: '#/components/schemas/MigrationPhase'
          description: Migration phases
        dependencies_graph:
          type: object
          additionalProperties: true
          description: Resource dependency graph
        migration_waves:
          type: array
          items:
            type: object
          description: Migration waves
        success_criteria:
          type: array
          items:
            type: string
          description: Success criteria
        created_at:
          type: string
          format: date-time
          description: Plan creation timestamp
      required:
        - plan_id
        - target_provider
        - total_duration_days
        - estimated_cost
        - risk_level
        - phases
    MigrationPhase:
      type: object
      properties:
        phase_id:
          type: string
          description: Unique phase identifier
        phase_name:
          type: string
          description: Phase name
        phase_order:
          type: integer
          description: Phase execution order
        workloads:
          type: array
          items:
            type: string
          description: Workloads in this phase
        start_date:
          type: string
          format: date-time
          description: Planned start date
        end_date:
          type: string
          format: date-time
          description: Planned end date
        actual_start_date:
          type: string
          format: date-time
          description: Actual start date
        actual_end_date:
          type: string
          format: date-time
          description: Actual end date
        status:
          type: string
          enum: [not_started, in_progress, completed, failed, blocked]
          description: Phase status
        prerequisites:
          type: array
          items:
            type: string
          description: Prerequisites for this phase
        success_criteria:
          type: array
          items:
            type: string
          description: Success criteria for this phase
        notes:
          type: string
          description: Additional notes
      required:
        - phase_id
        - phase_name
        - phase_order
        - status
...
security:
  - BearerAuth: []
paths:
...
  /api/migrations/{project_id}/plan:
    post:
      tags:
        - Migration Planning
      summary: Generate migration plan
      description: |
        Generate a comprehensive migration plan with phases, timelines, and cost estimates.
        
        **Requirements:** 4.1
      operationId: generateMigrationPlan
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_provider:
                  type: string
                  description: Target cloud provider
                workload_ids:
                  type: array
                  items:
                    type: string
                  description: List of workload IDs to migrate
                migration_strategy:
                  type: string
                  default: "phased"
                  enum: [phased, big_bang, parallel]
                  description: Migration strategy
                target_start_date:
                  type: string
                  format: date-time
                  description: Target migration start date
              required:
                - target_provider
            example:
              target_provider: "aws"
              workload_ids: ["workload-1", "workload-2"]
              migration_strategy: "phased"
              target_start_date: "2024-01-15T00:00:00"
      responses:
        '201':
          description: Migration plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          
    get:
      tags:
        - Migration Planning
      summary: Get migration plan
      description: |
        Retrieve the migration plan for a project.
        
        **Requirements:** 4.1
      operationId: getMigrationPlan
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      responses:
        '200':
          description: Migration plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
        '404':
          description: Plan not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/migrations/{project_id}/plan/progress:
    get:
      tags:
        - Migration Planning
      summary: Get migration progress
      description: |
        Get comprehensive migration progress tracking information.
        
        **Requirements:** 4.4
      operationId: getMigrationProgress
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      responses:
        '200':
          description: Migration progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan_id:
                    type: string
                  overall_status:
                    type: string
                  total_phases:
                    type: integer
                  completed_phases:
                    type: integer
                  in_progress_phases:
                    type: integer
                  not_started_phases:
                    type: integer
                  failed_phases:
                    type: integer
                  overall_progress_percentage:
                    type: number
                    format: float
                  current_phase:
                    $ref: '#/components/schemas/MigrationPhase'
                  next_phase:
                    $ref: '#/components/schemas/MigrationPhase'
                  estimated_completion_date:
                    type: string
                    format: date-time
                  actual_start_date:
                    type: string
                    format: date-time
                  days_elapsed:
                    type: integer
                  days_remaining:
                    type: integer
        '404':
          description: Plan not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...